+ finished poisson disc sampling where the minimum distance is based on 3d perlin noise

+++ GAW-Unity/Assets/Scripts/Helper.cs
@@ -21,7 +21,7 @@ public static class Helper
 public class MathS
 {
 
-	public static float FloorToInt(float value)
+	public static int FloorToInt(float value)
 	{
 		return (int)System.Math.Floor(value);
 	}
